Equipping Weapons
I am currently working on equipping weapons. I have gotten a new model that doesn't have the gun attached to it. I have to still set up the spawn point for the weapon on the model. I also need to instantiate the prefab. But I got the weapons models and the new character.